+++ python/pathway/io/_utils.py
@@ -49,6 +49,7 @@ _PATHWAY_TYPE_MAPPING: dict[PathwayType, dt.DType] = {
     PathwayType.JSON: dt.JSON,
     PathwayType.TUPLE: dt.ANY_TUPLE,
     PathwayType.BYTES: dt.BYTES,
+    PathwayType.PY_OBJECT_WRAPPER: dt.ANY_PY_OBJECT_WRAPPER,
 }
 
 SUPPORTED_INPUT_FORMATS: set[str] = {